fix(test_cache): Correct XDG testing logic
The tests on Linux were relying on XDG_CACHE_HOME being set to its
fallback value of ∼/.cache, which made them fail on systems with custom
values for XDG_CACHE_HOME.
To account for this, we take the cue from the Windows tests and set
XDG_CACHE_HOME to a known, custom, value. This incidentally also tests
that our logic can account for custom values of XDG_CACHE_HOME, so this
doesn't need its own test.
The value of "/tmp/home/.cache" was chosen as a cross between the
Windows value of "/tmp/AppData/Local" and the fallback "∼/.cache"

Fixes: #814